BeatDetection/requirements.txt
librosa
cython
mido
onnxruntime-directml